moved number format together
