+++ src/app/numberFormat/formatNumber.js
